# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Indiana Choice Scholarship Program Annual Report API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  username?: string;
  firstName?: string;
  lastName?: string;
  profileImage?: string;
  isAdmin: boolean;
}

export interface GetCurrentAuthUserResponse {
  isAuthenticated: boolean;
  user?: AuthUser;
}

export interface ReportSection {
  id: number;
  key: string;
  title: string;
  content: string;
  sectionOrder: number;
  parentKey?: string;
  updatedAt?: string;
}

export interface ReportTable {
  id: number;
  key: string;
  title: string;
  description?: string;
  headers: string[];
  rows: string[][];
  footnote?: string;
  sectionKey: string;
  tableOrder: number;
  updatedAt?: string;
}

export interface ReportStat {
  id: number;
  key: string;
  label: string;
  value: string;
  description?: string;
  sectionKey: string;
  statOrder: number;
  updatedAt?: string;
}

export interface ReportMetadata {
  id: number;
  reportTitle: string;
  reportYear: string;
  reportDescription?: string;
  publishedDate?: string;
  updatedAt?: string;
}

export interface AdminUser {
  id: number;
  replitUserId: string;
  username: string;
  addedAt: string;
}

export interface UpdateReportSectionBody {
  title: string;
  content: string;
}

export interface UpdateReportTableBody {
  title: string;
  description?: string;
  headers: string[];
  rows: string[][];
  footnote?: string;
}

export interface UpdateReportStatBody {
  label: string;
  value: string;
  description?: string;
}

export interface UpdateReportMetadataBody {
  reportTitle: string;
  reportYear: string;
  reportDescription?: string;
  publishedDate?: string;
}

export interface AddAdminUserBody {
  replitUserId: string;
  username: string;
}

export interface SuccessResponse {
  success: boolean;
}