# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Indiana Choice Scholarship Program Annual Report API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get current user
 */
export const GetCurrentAuthUserResponse = zod.object({
  isAuthenticated: zod.boolean(),
  user: zod
    .object({
      id: zod.string(),
      username: zod.string().optional(),
      firstName: zod.string().optional(),
      lastName: zod.string().optional(),
      profileImage: zod.string().optional(),
      isAdmin: zod.boolean(),
    })
    .optional(),
});

/**
 * @summary Get all report sections
 */
export const GetReportSectionsResponseItem = zod.object({
  id: zod.number(),
  key: zod.string(),
  title: zod.string(),
  content: zod.string(),
  sectionOrder: zod.number(),
  parentKey: zod.string().optional(),
  updatedAt: zod.coerce.date().optional(),
});
export const GetReportSectionsResponse = zod.array(
  GetReportSectionsResponseItem,
);

/**
 * @summary Get a report section by key
 */
export const GetReportSectionParams = zod.object({
  key: zod.coerce.string(),
});

export const GetReportSectionResponse = zod.object({
  id: zod.number(),
  key: zod.string(),
  title: zod.string(),
  content: zod.string(),
  sectionOrder: zod.number(),
  parentKey: zod.string().optional(),
  updatedAt: zod.coerce.date().optional(),
});

/**
 * @summary Get all report tables
 */
export const GetReportTablesResponseItem = zod.object({
  id: zod.number(),
  key: zod.string(),
  title: zod.string(),
  description: zod.string().optional(),
  headers: zod.array(zod.string()),
  rows: zod.array(zod.array(zod.string())),
  footnote: zod.string().optional(),
  sectionKey: zod.string(),
  tableOrder: zod.number(),
  updatedAt: zod.coerce.date().optional(),
});
export const GetReportTablesResponse = zod.array(GetReportTablesResponseItem);

/**
 * @summary Get a report table by key
 */
export const GetReportTableParams = zod.object({
  key: zod.coerce.string(),
});

export const GetReportTableResponse = zod.object({
  id: zod.number(),
  key: zod.string(),
  title: zod.string(),
  description: zod.string().optional(),
  headers: zod.array(zod.string()),
  rows: zod.array(zod.array(zod.string())),
  footnote: zod.string().optional(),
  sectionKey: zod.string(),
  tableOrder: zod.number(),
  updatedAt: zod.coerce.date().optional(),
});

/**
 * @summary Get all key statistics
 */
export const GetReportStatsResponseItem = zod.object({
  id: zod.number(),
  key: zod.string(),
  label: zod.string(),
  value: zod.string(),
  description: zod.string().optional(),
  sectionKey: zod.string(),
  statOrder: zod.number(),
  updatedAt: zod.coerce.date().optional(),
});
export const GetReportStatsResponse = zod.array(GetReportStatsResponseItem);

/**
 * @summary Get report metadata (title, year, description)
 */
export const GetReportMetadataResponse = zod.object({
  id: zod.number(),
  reportTitle: zod.string(),
  reportYear: zod.string(),
  reportDescription: zod.string().optional(),
  publishedDate: zod.string().optional(),
  updatedAt: zod.coerce.date().optional(),
});

/**
 * @summary Update a report section
 */
export const UpdateReportSectionParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateReportSectionBody = zod.object({
  title: zod.string(),
  content: zod.string(),
});

export const UpdateReportSectionResponse = zod.object({
  id: zod.number(),
  key: zod.string(),
  title: zod.string(),
  content: zod.string(),
  sectionOrder: zod.number(),
  parentKey: zod.string().optional(),
  updatedAt: zod.coerce.date().optional(),
});

/**
 * @summary Update a report table
 */
export const UpdateReportTableParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateReportTableBody = zod.object({
  title: zod.string(),
  description: zod.string().optional(),
  headers: zod.array(zod.string()),
  rows: zod.array(zod.array(zod.string())),
  footnote: zod.string().optional(),
});

export const UpdateReportTableResponse = zod.object({
  id: zod.number(),
  key: zod.string(),
  title: zod.string(),
  description: zod.string().optional(),
  headers: zod.array(zod.string()),
  rows: zod.array(zod.array(zod.string())),
  footnote: zod.string().optional(),
  sectionKey: zod.string(),
  tableOrder: zod.number(),
  updatedAt: zod.coerce.date().optional(),
});

/**
 * @summary Update a report statistic
 */
export const UpdateReportStatParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateReportStatBody = zod.object({
  label: zod.string(),
  value: zod.string(),
  description: zod.string().optional(),
});

export const UpdateReportStatResponse = zod.object({
  id: zod.number(),
  key: zod.string(),
  label: zod.string(),
  value: zod.string(),
  description: zod.string().optional(),
  sectionKey: zod.string(),
  statOrder: zod.number(),
  updatedAt: zod.coerce.date().optional(),
});

/**
 * @summary Update report metadata
 */
export const UpdateReportMetadataBody = zod.object({
  reportTitle: zod.string(),
  reportYear: zod.string(),
  reportDescription: zod.string().optional(),
  publishedDate: zod.string().optional(),
});

export const UpdateReportMetadataResponse = zod.object({
  id: zod.number(),
  reportTitle: zod.string(),
  reportYear: zod.string(),
  reportDescription: zod.string().optional(),
  publishedDate: zod.string().optional(),
  updatedAt: zod.coerce.date().optional(),
});

/**
 * @summary List all admin users
 */
export const GetAdminUsersResponseItem = zod.object({
  id: zod.number(),
  replitUserId: zod.string(),
  username: zod.string(),
  addedAt: zod.coerce.date(),
});
export const GetAdminUsersResponse = zod.array(GetAdminUsersResponseItem);

/**
 * @summary Add a new admin user
 */
export const AddAdminUserBody = zod.object({
  replitUserId: zod.string(),
  username: zod.string(),
});

/**
 * @summary Remove an admin user
 */
export const RemoveAdminUserParams = zod.object({
  id: zod.coerce.number(),
});

export const RemoveAdminUserResponse = zod.object({
  success: zod.boolean(),
});

export const ExchangeMobileAuthorizationCodeBody = zod.object({
  code: zod.string(),
  code_verifier: zod.string(),
  redirect_uri: zod.string(),
  state: zod.string(),
  nonce: zod.string().optional(),
});

export const ExchangeMobileAuthorizationCodeResponse = zod.object({
  token: zod.string(),
});

export const LogoutMobileSessionResponse = zod.object({
  success: zod.boolean(),
});