# lib/api-zod/src/generated/types/addAdminUserBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Indiana Choice Scholarship Program Annual Report API
 * OpenAPI spec version: 0.1.0
 */

export interface AddAdminUserBody {
  replitUserId: string;
  username: string;
}