# lib/api-zod/src/generated/types/reportSection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Indiana Choice Scholarship Program Annual Report API
 * OpenAPI spec version: 0.1.0
 */

export interface ReportSection {
  id: number;
  key: string;
  title: string;
  content: string;
  sectionOrder: number;
  parentKey?: string;
  updatedAt?: Date;
}