# lib/api-zod/src/generated/types/reportStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Indiana Choice Scholarship Program Annual Report API
 * OpenAPI spec version: 0.1.0
 */

export interface ReportStat {
  id: number;
  key: string;
  label: string;
  value: string;
  description?: string;
  sectionKey: string;
  statOrder: number;
  updatedAt?: Date;
}