# lib/api-zod/src/generated/types/reportTable.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Indiana Choice Scholarship Program Annual Report API
 * OpenAPI spec version: 0.1.0
 */

export interface ReportTable {
  id: number;
  key: string;
  title: string;
  description?: string;
  headers: string[];
  rows: string[][];
  footnote?: string;
  sectionKey: string;
  tableOrder: number;
  updatedAt?: Date;
}